requirements.txt
ultralytics
opencv-python
numpy
pillow
streamlit
rembg
pywavelets
torch
torchvision
